+++ src/Rule/HaveIBeenPwnedRule.php
@@ -155,7 +155,7 @@ final class HaveIBeenPwnedRule implements Rule
     /**
      * @param string $password Password to check in breaches.
      * @return int Number of appearances in breaches.
-     * @throws RuleException If an error occurred.
+     * @throws CouldNotUseRuleException If an error occurred.
      */
     private function getCount(string $password): int
     {
@@ -180,7 +180,7 @@ final class HaveIBeenPwnedRule implements Rule
             }
             return 0;
         } catch (ClientExceptionInterface | RuntimeException $exception) {
-            throw new RuleException(
+            throw new CouldNotUseRuleException(
                 $this,
                 'An error occurred while using the Have I Been Pwned? service: '.$exception->getMessage(),
                 $exception