+++ tests/Rule/HaveIBeenPwnedRuleTest.php
@@ -232,7 +232,7 @@ final class HaveIBeenPwnedRuleTest extends TestCase
         $rule = new HaveIBeenPwnedRule(5, 0);
         $rule->setClient($client);
 
-        $this->expectException(RuleException::class);
+        $this->expectException(CouldNotUseRuleException::class);
 
         $rule->test('291vnnzrvtu9');
     }